enemy anim updated
