+++ Assets/Animation/Enemy/StaticBoom/Alert.anim
@@ -21,9 +21,9 @@ AnimationClip:
   - curve:
     - time: 0
       value: {fileID: -1211587276, guid: f8377225fac0bdf41b44037988efeb5d, type: 3}
-    - time: 0.25
-      value: {fileID: -1615817803, guid: f8377225fac0bdf41b44037988efeb5d, type: 3}
     - time: 0.5
+      value: {fileID: -1615817803, guid: f8377225fac0bdf41b44037988efeb5d, type: 3}
+    - time: 1
       value: {fileID: -1615817803, guid: f8377225fac0bdf41b44037988efeb5d, type: 3}
     attribute: m_Sprite
     path: 
@@ -52,7 +52,7 @@ AnimationClip:
     m_AdditiveReferencePoseClip: {fileID: 0}
     m_AdditiveReferencePoseTime: 0
     m_StartTime: 0
-    m_StopTime: 0.51666665
+    m_StopTime: 1.0166667
     m_OrientationOffsetY: 0
     m_Level: 0
     m_CycleOffset: 0
