+++ Assets/Prefabs/Enemy/Flying-Bullet.prefab
@@ -28,7 +28,7 @@ Transform:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 2343930456232581135}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 1.7, y: 0.181, z: 0}
+  m_LocalPosition: {x: 1.68, y: 0.7, z: 0}
   m_LocalScale: {x: 2, y: 2, z: 1}
   m_ConstrainProportionsScale: 0
   m_Children: []
@@ -100,7 +100,7 @@ Animator:
   m_Controller: {fileID: 9100000, guid: b6e5f10a061bc2544b655a5486cdba63, type: 2}
   m_CullingMode: 0
   m_UpdateMode: 0
-  m_ApplyRootMotion: 0
+  m_ApplyRootMotion: 1
   m_LinearVelocityBlending: 0
   m_StabilizeFeet: 0
   m_WarningMessage: 
@@ -120,13 +120,13 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: b38eb2f556a202d4d9b66985670fe97b, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  enemyState: {fileID: 0}
   enemyEnergy: 0
   attackScale: 0
   currentInterval: 0
   maxInterval: 0
   player: {fileID: 6728937369562067951, guid: b5772a3044cba164abf87f127f1fee16, type: 3}
   animator: {fileID: 0}
-  bulletPrefab: {fileID: 1405041303486870691, guid: f0c3ccd3ebb18c04b90cd5a32d52e0bd, type: 3}
 --- !u!58 &4867353768541422633
 CircleCollider2D:
   m_ObjectHideFlags: 0
