+++ Assets/Scripts/Core/State.cs
@@ -46,29 +46,5 @@ namespace GbitProjectState
 			death = false;
 		}
 	}
-    public class EnemyState
-    {
-
-        public enum State
-        {
-            Idle,
-            Attack,
-			Alert,
-            Dead
-        }
-
-        private State currentState;
-
-        public State CurrentState
-        {
-            get { return currentState; }
-            set { currentState = value; }
-        }
-
-        public EnemyState()
-        {
-            currentState = State.Idle;
-        }
-    }
 }
